backend/requirements.txt
Flask
gunicorn
pylint
paho-mqtt
peewee
mariadb==1.0.11
APScheduler